chore: refactor projet structure
