+++ src/armonik_cli/commands/__init__.py
@@ -1,0 +1,4 @@
+from .sessions import sessions
+
+
+__all__ = ["sessions"]
